# speed up FT0 simulation