--- a/Detectors/FIT/FT0/simulation/include/FT0Simulation/Detector.h
+++ b/Detectors/FIT/FT0/simulation/include/FT0Simulation/Detector.h
@@ -148,7 +148,11 @@ class Detector : public o2::base::DetImpl<Detector>
 
   template <typename Det>
   friend class o2::base::DetImpl;
-  ClassDefOverride(Detector, 1);
+
+  int mTrackIdTop;
+  int mTrackIdMCPtop; //TEMPORARY
+
+  ClassDefOverride(Detector, 2);
 };
 
 // Input and output function for standard C++ input/output.