--- a/Detectors/FIT/FT0/simulation/src/Detector.cxx
+++ b/Detectors/FIT/FT0/simulation/src/Detector.cxx
@@ -34,7 +34,7 @@ using o2::ft0::Geometry;
 ClassImp(Detector);
 
 Detector::Detector(Bool_t Active)
-  : o2::base::DetImpl<Detector>("FT0", Active), mIdSens1(0), mPMTeff(nullptr), mHits(o2::utils::createSimVector<o2::ft0::HitType>())
+  : o2::base::DetImpl<Detector>("FT0", Active), mIdSens1(0), mPMTeff(nullptr), mHits(o2::utils::createSimVector<o2::ft0::HitType>()), mTrackIdTop(-1), mTrackIdMCPtop(-1)
 
 {
   // Gegeo  = GetGeometry() ;
@@ -43,7 +43,7 @@ Detector::Detector(Bool_t Active)
 }
 
 Detector::Detector(const Detector& rhs)
-  : o2::base::DetImpl<Detector>(rhs), mIdSens1(rhs.mIdSens1), mPMTeff(rhs.mPMTeff), mHits(o2::utils::createSimVector<o2::ft0::HitType>())
+  : o2::base::DetImpl<Detector>(rhs), mIdSens1(rhs.mIdSens1), mPMTeff(rhs.mPMTeff), mHits(o2::utils::createSimVector<o2::ft0::HitType>()), mTrackIdTop(-1), mTrackIdMCPtop(-1)
 {
 }
 
@@ -61,6 +61,19 @@ void Detector::InitializeO2Detector()
   } else {
 
     AddSensitiveVolume(v);
+  }
+
+  TGeoVolume* vrad = gGeoManager->GetVolume("0TOP");
+  if (vrad == nullptr) {
+    LOG(WARN) << "@@@@ Sensitive radiator not found!!!!!!!!";
+  } else {
+    AddSensitiveVolume(vrad);
+  }
+  TGeoVolume* vmcp = gGeoManager->GetVolume("0MTO");
+  if (vmcp == nullptr) {
+    LOG(WARN) << "@@@@ Sensitive MCP glass not found!!!!!!!!";
+  } else {
+    AddSensitiveVolume(vmcp);
   }
 }
 
@@ -233,7 +246,7 @@ void Detector::SetOneMCP(TGeoVolume* ins)
   Float_t pmcptopglass[3] = {2.949, 2.949, 0.1}; // MCP top glass optical
 
   Float_t preg[3] = {1.324, 1.324, 0.005}; // Photcathode
-  Double_t pal[3] = {2.648, 2.648, 0.25}; // 5mm Al on top of each radiator
+  Double_t pal[3] = {2.648, 2.648, 0.25};  // 5mm Al on top of each radiator
   // Entry window (glass)
   TVirtualMC::GetMC()->Gsvolu("0TOP", "BOX", getMediumID(kOpGlass), ptop, 3); // Glass radiator
   TGeoVolume* top = gGeoManager->GetVolume("0TOP");
@@ -326,6 +339,9 @@ void Detector::SetOneMCP(TGeoVolume* ins)
 
 Bool_t Detector::ProcessHits(FairVolume* v)
 {
+
+  TString volname = fMC->CurrentVolName();
+
   TVirtualMCStack* stack = fMC->GetStack();
   Int_t quadrant, mcp;
   if (fMC->IsTrackEntering()) {
@@ -343,10 +359,39 @@ Bool_t Detector::ProcessHits(FairVolume* v)
     if (fMC->TrackCharge()) { //charge particles for MCtrue
       AddHit(x, y, z, time, 10, trackID, detID);
     }
-    if (iPart == 50000050) // If particles is photon then ...
-    {
-      if (RegisterPhotoE(etot)) {
-        AddHit(x, y, z, time, enDep, parentID, detID);
+    if (iPart == 50000050) { // If particles is photon then ...
+      if (etot > 7.46e-9 || etot < 2.4e-9) {
+        fMC->StopTrack();
+        return kFALSE;
+      }
+      if (volname.Contains("0TOP")) {
+        if (!RegisterPhotoE(etot)) {
+          fMC->StopTrack();
+          return kFALSE;
+        }
+        mTrackIdTop = trackID;
+      }
+
+      if (volname.Contains("0MTO")) {
+        if (trackID != mTrackIdTop) {
+          if (!RegisterPhotoE(etot)) {
+            //   std::cout<<" брысь "<<etot<<" track "<<trackID<<" parentID "<<parentID<<" "<<volname<<std::endl;
+            fMC->StopTrack();
+            return kFALSE;
+          }
+          mTrackIdMCPtop = trackID;
+        }
+      }
+
+      if (volname.Contains("0REG")) {
+        if (trackID != mTrackIdTop && trackID != mTrackIdMCPtop) {
+          if (RegisterPhotoE(etot)) {
+            AddHit(x, y, z, time, enDep, parentID, detID);
+          }
+        }
+        if (trackID == mTrackIdTop || trackID == mTrackIdMCPtop) {
+          AddHit(x, y, z, time, enDep, parentID, detID);
+        }
       }
     }
 
@@ -358,8 +403,10 @@ Bool_t Detector::ProcessHits(FairVolume* v)
 o2::ft0::HitType* Detector::AddHit(float x, float y, float z, float time, float energy, Int_t trackId, Int_t detId)
 {
   mHits->emplace_back(x, y, z, time, energy, trackId, detId);
-  auto stack = (o2::data::Stack*)fMC->GetStack();
-  stack->addHit(GetDetId());
+  if (energy == 10) {
+    auto stack = (o2::data::Stack*)fMC->GetStack();
+    stack->addHit(GetDetId());
+  }
   return &(mHits->back());
 }
 
@@ -424,7 +471,7 @@ void Detector::CreateMaterials()
   Medium(18, "OpBlack$", 2, 0, isxfld, sxmgmx, 10., .1, 1., .003, .003);
   Medium(15, "Aluminium$", 11, 0, isxfld, sxmgmx, 10., .01, 1., .003, .003);
   Medium(17, "OptAluminium$", 11, 0, isxfld, sxmgmx, 10., .01, 1., .003, .003);
-  Medium(16, "OpticalGlass$", 24, 1, isxfld, sxmgmx, 10., .01, .1, .003, .003);
+  Medium(16, "OpticalGlass$", 24, 1, isxfld, sxmgmx, 10., .01, .1, .003, .01);
   Medium(19, "OpticalGlassCathode$", 24, 1, isxfld, sxmgmx, 10., .01, .1, .003, .003);
   Medium(22, "SensAir$", 2, 1, isxfld, sxmgmx, 10., .1, 1., .003, .003);
 }