- increased min step size in thick radiator and front quartz MCP window;
- added radiator (0TOP) and front quartz MCP window (0MOP) as sensitive volumes;
- when photons just appeared we check QE and stop transport of photon if not;
- stop writing Cherenkov photons in Kine tree.  

CPU time was dropped down

- for pp from 0.43 to 0.25 s / event;
- for Pb-Pb  from 29 to 17s / event.