@sawenzel 2020-05-28
Very nice! I will merge this but like to finish a step analysis in the current state to see the impact of this PR.